+++ samigo/samigo-api/src/java/org/sakaiproject/tool/assessment/shared/api/questionpool/QuestionPoolServiceAPI.java
@@ -74,7 +74,7 @@ public interface QuestionPoolServiceAPI
   /**
    * Get the size of a subpool.
    */
-  public int getSubPoolSize(Long poolId);
+  public long getSubPoolSize(Long poolId);
 
   /**
    * Checks to see if a pool has subpools
